fetch decode execute done

+++ src/DataMemory.java
@@ -1,11 +1,11 @@
 public class DataMemory {
-    private static int[] memory = new int[2048]; //each row is 8 bits (1 int)
+    private static byte[] memory = new byte[2048]; //each row is 8 bits (1 byte)
 
-    public static int[] getMemory() {
+    public static byte[] getMemory() {
         return memory;
     }
 
-    public static void setMemory(int address, int value) {
+    public static void setMemory(int address, byte value) {
         DataMemory.memory[address] = value;
     }
 }
